+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This CHANGELOG follows the format listed at [Our CHANGELOG Guidelines ](https://
 Which is based on [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+### Changed
+- bin/metrics-*: Used `Sensu::Plugin::Metric::CLI::Generic` class instead of Graphite specific class for metrics.
 
 ## [2.4.1] - 2018-01-23
 ### Fixed